Add files via upload
Homework was about "Choosing an existing website" So I went with the 'Amazon Store' website. this HW was in (SE-2111) Software Engineering,
I hope you enjoy reading the research and be beneficial to you.

Finally, I would like to thank Dr.Mohammed Assiri very much for giving me the opportunity to share my experiences on the site and for all it has provided us. 
Wishing him continued success